requirements.txt
streamlit
Pillow
groq
pandas
numpy
scikit-learn
sentence-transformers
serpapi
